Move logic inside RestrictedAccessBody + allow members in projects
